Edit BaseViewModel and move MainViewModel
Took 24 minutes
