+++ DentalCore.Wpf/Abstract/BaseViewModel.cs
@@ -1,10 +1,11 @@
-﻿using System.ComponentModel;
+﻿using System;
+using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Windows.Input;
 
 namespace DentalCore.Wpf.Abstract;
 
-public abstract class BaseViewModel : INotifyPropertyChanged
+public abstract class BaseViewModel : INotifyPropertyChanged, IDisposable
 {
     public event PropertyChangedEventHandler? PropertyChanged;
     public ICommand LoadedCommand { get; protected init; }
